// Client/src/models/EmployeeDto.tsx
export interface EmployeeDto {
  id: string;
  firstName: string;
  lastName: string;
  dateOfBirth: Date;
  jobTitle: string;
  imageUrl: string;
  email: string;
  homeAddress: string;
  postalCode: string;
  city: string;
  bankAccountNumber: string;
  taxRate: number;
  grossSalary: number;
  startDate: Date;
}